Fix: ignorePaths option not matching correctly
Passing dirent.name to ignorePaths matcher will not correctly match all the times. This is evident when we want to ignore any file that starts with 'index*' only in src folder and with previous implementation, all file that matches this pattern at any depth will be ignored. So, fix this by testing the complete path (both relative and absolute) against the matcher. Now, we have to start with '**/' to match any depth and eventually we can restrict only to src folder by specifying the pattern as '**/src/index*'. That '**' is needed as we are matching against the complete path.