Feat: add withFilePath option
Add feature to get the list of files without path prepended to it by setting the option with false. Also, make flatten 'true' no effect when withFilePath is 'false'. Because, we don't want to have the same filename from different dir end up in the final array, leave it that to the user.